requirements.txt
finnhub-python==2.4.19
python-dotenv==1.0.0
websocket-client==1.7.0
confluent-kafka==2.3.0
pytest==7.4.4
pre-commit==3.6.0
ruff==0.1.11
pytest-cov==4.1.0
